Fixed bStop on programming vias
There is bStop added to expose the bottom side of the programming vias.
It used to be a circle, which did not expose 100% of the via - it still
contained a little "dot" of mask covering up the center of the vias.
This has made it difficult (if not impossible) to program.  On this
commit, I added a circle drawn with the polygon tool to the bStop on all
6 vias to expose 100% of the vias.